requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
audioop-lts==0.2.1
    # via
    #   -r requirements.in
    #   standard-aifc
    #   standard-sunau
audioread==3.0.1
    # via
    #   -r requirements.in
    #   librosa
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2025.4.26
    # via
    #   -r requirements.in
    #   requests
cffi==1.17.1
    # via
    #   -r requirements.in
    #   soundfile
charset-normalizer==3.4.2
    # via
    #   -r requirements.in
    #   requests
click==8.2.1
    # via
    #   -r requirements.in
    #   pip-tools
decorator==5.2.1
    # via
    #   -r requirements.in
    #   librosa
ffmpeg-python==0.2.0
    # via -r requirements.in
future==1.0.0
    # via
    #   -r requirements.in
    #   ffmpeg-python
idna==3.10
    # via
    #   -r requirements.in
    #   requests
joblib==1.5.1
    # via
    #   -r requirements.in
    #   librosa
    #   scikit-learn
lazy-loader==0.4
    # via
    #   -r requirements.in
    #   librosa
librosa==0.11.0
    # via -r requirements.in
llvmlite==0.44.0
    # via
    #   -r requirements.in
    #   numba
msgpack==1.1.0
    # via
    #   -r requirements.in
    #   librosa
numba==0.61.2
    # via
    #   -r requirements.in
    #   librosa
numpy==2.2.6
    # via
    #   librosa
    #   numba
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
packaging==25.0
    # via
    #   -r requirements.in
    #   build
    #   lazy-loader
    #   pooch
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.8
    # via
    #   -r requirements.in
    #   pooch
pooch==1.8.2
    # via
    #   -r requirements.in
    #   librosa
pyaudio==0.2.14
    # via -r requirements.in
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pydub==0.25.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
requests==2.32.3
    # via
    #   -r requirements.in
    #   pooch
ruff==0.11.13
    # via -r requirements.in
scikit-learn==1.7.0
    # via
    #   -r requirements.in
    #   librosa
scipy==1.15.3
    # via
    #   -r requirements.in
    #   librosa
    #   scikit-learn
soundfile==0.13.1
    # via
    #   -r requirements.in
    #   librosa
soxr==0.5.0.post1
    # via
    #   -r requirements.in
    #   librosa
standard-aifc==3.13.0
    # via
    #   -r requirements.in
    #   librosa
standard-chunk==3.13.0
    # via
    #   -r requirements.in
    #   standard-aifc
standard-sunau==3.13.0
    # via
    #   -r requirements.in
    #   librosa
threadpoolctl==3.6.0
    # via
    #   -r requirements.in
    #   scikit-learn
typing-extensions==4.14.0
    # via
    #   -r requirements.in
    #   librosa
urllib3==2.4.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools
yt-dlp==2025.5.22
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
